requirements.txt
langchain==0.2.12
langchain-openai==0.1.21
langgraph==0.2.3
python-dotenv==1.0.1
pandas==2.2.2